Seperated the PollEntry Dialog from the Voting Module.
