+++ WarBot.Modules/WarBot.Modules.csproj
@@ -13,14 +13,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Remove="Dialogs\PollOptionEntryDialog.cs" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Include="Dialogs\PollOptionEntryDialog.cs" />
-  </ItemGroup>
-
-  <ItemGroup>
     <PackageReference Include="Humanizer.Core" Version="2.5.1" />
   </ItemGroup>
